refactor(upload): move methods between api and watermark/canvas modules
